youtube_agent/requirements.txt
# Google APIs
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.130.0
google-generativeai==0.7.2

# Telegram
python-telegram-bot==21.3

# Automation (Phase 5)
playwright==1.44.0

# Utilities
python-dotenv==1.0.1
schedule==1.2.2
pytz==2024.1
aiohttp==3.9.5

# Dev / Testing
pytest==8.2.2
